Refactoring Http Request With Dependency Injection
