+++ APIBagLib/APICountryLib.csproj
@@ -8,4 +8,8 @@
     <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ApiBase\ApiBase.csproj" />
+  </ItemGroup>
+
 </Project>
